[PH-186] 로그인 후에 계정페이지 등 안들어가지는 현상 수정
