+++ src/components/auth/Login.tsx
@@ -1,24 +1,15 @@
 "use client";
 
 import Link from "next/link";
-import { useRouter } from "next/navigation";
 import React, { useState } from "react";
 import { FaEye, FaEyeSlash } from "react-icons/fa";
 import { TiDelete } from "react-icons/ti";
 import { login } from "@/services/auth.service";
-import { setIsAuthenticated } from "@/stores/authSlice";
-import { useDispatch } from "react-redux";
-import { getUser } from "@/services/user.service";
-import { setUser } from "@/stores/userSlice";
 
 const Login = () => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [isVisivlePassword, setIsVisivlePassword] = useState(false);
-
-  const router = useRouter();
-
-  const dispatch = useDispatch();
 
   const handleEmailChange = (event: React.ChangeEvent<HTMLInputElement>) => {
     setEmail(event.target.value);
@@ -46,22 +37,7 @@ const Login = () => {
 
     const response = await login({ email, password });
     if (response) {
-      dispatch(setIsAuthenticated(true));
-      const fetchUser = async () => {
-        const data = await getUser();
-        if (data) {
-          dispatch(
-            setUser({
-              email: data.email,
-              nickname: data.nickname,
-              bio: data.bio,
-              profileImage: data.image,
-            }),
-          );
-        }
-      };
-      fetchUser();
-      router.replace("/");
+      window.location.href = "/";
     }
   };
